Simplify sidebar as static HTML
This commit simplifies the sidebar in index.html by converting it from
a JavaScript component to static HTML markup. The sidebar is a simple
layout component, and using JavaScript for this was unnecessary and
overcomplicated.

The new static HTML sidebar includes a div with the id of
"project-list-component", which will serve as the entry point for a
dynamic component that will render the list of projects present in the
projects data. This change is part of an ongoing effort to simplify
the codebase and improve maintainability.