Add ProjectList component and render it in index.js
This commit introduces a new component, ProjectList, which is
responsible for displaying a list of projects.

The component uses the 'getProjects' method from the ProjectController
to fetch all the projects and creates a list item for each project.

This commit also updates 'src/index.js' to render the ProjectList
component. The ProjectList component is appended to the
'project-list-component' element in the DOM.

With the addition and rendering of this new component, users will
now be able to see a list of all projects in the application's UI.