Move project title from ProjectList to index.html
This commit modifies the ProjectList component to return only the list
of projects, without the title. The title "Projects" is now added
directly in the 'index.html' file, above the 'project-list-component'
element.

This change simplifies the ProjectList component and allows for better
separation of concerns, as the title is now part of the static HTML
structure and not dynamically created by the ProjectList component.