Fix component name in CreateProjectDialog
This commit corrects the component name in the CreateProjectDialog
component. Previously, the 'ProjectList()' function was used in the
refreshComponent call, which was a mistake. The correct component
name is 'ProjectListItems()'. This bug occurred due to an oversight
when updating the component name. This fix ensures that the correct
component is refreshed when a new project is created.