Introduce project deletion functionality
This commit introduces the ability to delete a project from the project
component. A new 'DeleteProjectBtn' function has been added which creates
a button that, when clicked, deletes the project with the given id and
refreshes the project list and project board components.

In addition, an 'Actions' function has been created which appends the
delete button to the project header. This provides users with an easy
and intuitive way to manage their projects.