Move utility functions to state manager
This commit moves the 'updateComponent' and 'initComponent' functions
from the 'utils.js' file in the 'components' directory to a new file
called 'state-manager.js' in the 'state' directory. This change is made
to improve the organization of the codebase, grouping state management
functions together in a dedicated file. This should make the code easier
to navigate and maintain in the future.