Refactor project and task creation to use Factory class
This commit replaces the individual project and task factories with a
unified Factory class. The Factory class contains methods for creating
projects and tasks, which are now invoked as 'factory.createProject' and
'factory.createTask'. This change makes it clearer in the application
when the factory methods are being called, as opposed to the similarly
named 'stateManager.createProject'.

The individual project and task factory files have been deleted, and all
instances of 'createProject' and 'createTask' have been updated to use
the new Factory methods. This refactoring helps to avoid confusion
between the factory methods and the state manager methods.